move import processing to the queue
